Isolate and namespace `CableReady::Installer` (#295)
## Description

This pull request isolates the methods defined
`lib/cable_ready/installer.rb` into a `CableReady::Installer` constant
so they don't pollute the global namespace if the file is required.

## Why should this be added

Exposing all the methods from the installer in the global namespace
could lead to issues down the road. Methods like `options` are likely to
be overridden.